Handled edge case in `mlo_smo_prep`: no new SMO data
There was no new SMO data in May 2024, so the run to add May 2024
data to the monthly file failed. It crashed when trying to apply
the wind filter because it had no dates from the in situ data to
use to figure out if we have all the necessary GEOS files. The
fix was to just skip that check, which is unecessary anyway if
no data is available.

This particular bug won't affect MLO because only SMO needs to
filted on wind direction.