Added program to download f(O2) inputs
Also moved `GeosVersion` and `GeosSource` so that I could put
the functions to calculate checksums in the `mod_utils` modules
without a possible circular import between `mod_utils` and `mod_constants`.

This was cherry-picked from the main branch to support adding O2 DMFs
to `.vmr` file headers in pre-April 2024 files.