Small fixes for O2 DMF plus update test files.
There were some changes to the readers that were needed to handle
the most recent format for GEOS filenames in the `.mod` file headers
and the `.map.nc` dates. Also, some of the O2 code itself was not
compatible with the earlier package versions used in this LTS version,
that has been fixed. The test files are now up to date with the correct
headers; the `.mod` and `.vmr` files changed in the headers only, the
`.map` files did not change at all.